hook up actions correctly
