+++ src/vs/workbench/contrib/chat/browser/chatEditing/notebook/overlayToolbarController.ts
@@ -1,0 +1,122 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Disposable, DisposableStore } from '../../../../../../base/common/lifecycle.js';
+import { AccessibilitySignal, IAccessibilitySignalService } from '../../../../../../platform/accessibilitySignal/browser/accessibilitySignalService.js';
+import { MenuWorkbenchToolBar, HiddenItemStrategy } from '../../../../../../platform/actions/browser/toolbar.js';
+import { MenuId } from '../../../../../../platform/actions/common/actions.js';
+import { IContextKeyService } from '../../../../../../platform/contextkey/common/contextkey.js';
+import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
+import { ServiceCollection } from '../../../../../../platform/instantiation/common/serviceCollection.js';
+import { INotebookEditor } from '../../../../notebook/browser/notebookBrowser.js';
+import { NotebookTextModel } from '../../../../notebook/common/model/notebookTextModel.js';
+import { IModifiedFileEntryChangeHunk } from '../../../common/chatEditingService.js';
+import { ICellDiffInfo } from './notebookCellChanges.js';
+
+
+export class OverlayToolbarController extends Disposable {
+
+	private _timeout: any | undefined = undefined;
+	private readonly overlayDisposables = this._register(new DisposableStore());
+
+	constructor(
+		private readonly notebookEditor: INotebookEditor,
+		private readonly notebookModel: NotebookTextModel,
+		@IInstantiationService private readonly instantiationService: IInstantiationService,
+		@IAccessibilitySignalService private readonly accessibilitySignalService: IAccessibilitySignalService,
+	) {
+		super();
+	}
+
+	decorate(changes: ICellDiffInfo[]) {
+		if (this._timeout !== undefined) {
+			clearTimeout(this._timeout);
+		}
+		this._timeout = setTimeout(() => {
+			this._timeout = undefined;
+			this.createMarkdownPreviewToolbars(changes);
+		}, 100);
+	}
+
+	private createMarkdownPreviewToolbars(changes: ICellDiffInfo[]) {
+		this.overlayDisposables.clear();
+
+		const accessibilitySignalService = this.accessibilitySignalService;
+		const editor = this.notebookEditor;
+		for (const change of changes) {
+			const cellViewModel = this.getCellViewModel(change);
+			if (!cellViewModel) {
+				continue;
+			}
+			editor.changeCellOverlays((accessor) => {
+				const toolbarContainer = document.createElement('div');
+				toolbarContainer.style.right = '44px';
+				const overlayId = accessor.addOverlay({
+					cell: cellViewModel,
+					domNode: toolbarContainer,
+				});
+
+				const toolbar = document.createElement('div');
+				toolbarContainer.appendChild(toolbar);
+				toolbar.className = 'chat-diff-change-content-widget';
+				toolbar.classList.add('hover'); // Show by default
+				toolbar.style.position = 'relative';
+				toolbar.style.top = '18px';
+				toolbar.style.zIndex = '10';
+
+				const removeOverlay = () => {
+					editor.changeCellOverlays(accessor => {
+						accessor.removeOverlay(overlayId);
+					});
+				};
+
+				this.overlayDisposables.add({ dispose: removeOverlay });
+
+				const scopedInstaService = this._register(this.instantiationService.createChild(new ServiceCollection([IContextKeyService, this.notebookEditor.scopedContextKeyService])));
+				const toolbarWidget = scopedInstaService.createInstance(MenuWorkbenchToolBar, toolbar, MenuId.ChatEditingEditorHunk, {
+					telemetrySource: 'chatEditingNotebookHunk',
+					hiddenItemStrategy: HiddenItemStrategy.NoHide,
+					toolbarOptions: { primaryGroup: () => true },
+					menuOptions: {
+						renderShortTitle: true,
+						arg: {
+							accept() {
+								accessibilitySignalService.playSignal(AccessibilitySignal.editsKept, { allowManyInParallel: true });
+								removeOverlay();
+								toolbarWidget.dispose();
+								return change.keep(change.diff.get().changes[0]);
+							},
+							reject() {
+								accessibilitySignalService.playSignal(AccessibilitySignal.editsUndone, { allowManyInParallel: true });
+								removeOverlay();
+								toolbarWidget.dispose();
+								return change.undo(change.diff.get().changes[0]);
+							}
+						} satisfies IModifiedFileEntryChangeHunk,
+					},
+				});
+
+				this.overlayDisposables.add(toolbarWidget);
+			});
+		}
+	}
+
+	private getCellViewModel(change: ICellDiffInfo) {
+		if (change.type === 'delete' || change.modifiedCellIndex === undefined) {
+			return undefined;
+		}
+		const cell = this.notebookModel.cells[change.modifiedCellIndex];
+		const cellViewModel = this.notebookEditor.getViewModel()?.viewCells.find(c => c.handle === cell.handle);
+		return cellViewModel;
+	}
+
+	override dispose(): void {
+		super.dispose();
+		if (this._timeout !== undefined) {
+			clearTimeout(this._timeout);
+		}
+	}
+
+}
